+++ addon/components/base-layer.js
@@ -1,13 +1,31 @@
 import Ember from 'ember';
-import ChildMixin from 'ember-leaflet/mixins/child';
+import { ChildMixin } from 'ember-composability-tools';
 import { InvokeActionMixin } from 'ember-invoke-action';
 /* global L */
 
-const { assert, computed, Component, run, K, A, String: { classify } } = Ember;
+const {
+  assert,
+  computed,
+  Component,
+  getOwner,
+  run,
+  A,
+  String: { classify }
+} = Ember;
+
+const leaf = typeof L === 'undefined' ? {} : L;
 
 export default Component.extend(ChildMixin, InvokeActionMixin, {
   tagName: '',
-  L,
+  L: leaf,
+
+  fastboot: computed(function() {
+    let owner = getOwner(this);
+    return owner.lookup('service:fastboot');
+  }),
+  isFastBoot: computed('fastboot', function() {
+    return this.get('fastboot') && this.get('fastboot.isFastBoot');
+  }),
 
   concatenatedProperties: ['leafletOptions', 'leafletRequiredOptions', 'leafletEvents', 'leafletProperties'],
 
@@ -15,17 +33,22 @@ export default Component.extend(ChildMixin, InvokeActionMixin, {
     assert('BaseLayer\'s `createLayer` should be overriden.');
   },
 
-  didCreateLayer: K,
-  willDestroyLayer: K,
+  didCreateLayer() {},
+  willDestroyLayer() {},
 
   /*
    * Method called by parent when the layer needs to setup
    */
-  layerSetup() {
+  didInsertParent() {
+    // Check for fastBoot
+    if (this.get('isFastBoot')) {
+      return;
+    }
+
     this._layer = this.createLayer();
     this._addObservers();
     this._addEventListeners();
-    if (this.get('containerLayer')) {
+    if (this.get('parentComponent')) {
       this.addToContainer();
     }
     this.didCreateLayer();
@@ -35,27 +58,32 @@ export default Component.extend(ChildMixin, InvokeActionMixin, {
    * Default logic for adding the layer to the container
    */
   addToContainer() {
-    this.get('containerLayer')._layer.addLayer(this._layer);
+    this.get('parentComponent')._layer.addLayer(this._layer);
   },
 
   /*
    * Method called by parent when the layer needs to teardown
    */
-  layerTeardown() {
+  willDestroyParent() {
+    // Check for fastBoot
+    if (this.get('isFastBoot')) {
+      return;
+    }
+
     this.willDestroyLayer();
     this._removeEventListeners();
     this._removeObservers();
-    if (this.get('containerLayer') && this._layer) {
+    if (this.get('parentComponent') && this._layer) {
       this.removeFromContainer();
     }
-    this._layer = null;
+    delete this._layer;
   },
 
   /*
    * Default logic for removing the layer from the container
    */
   removeFromContainer() {
-    this.get('containerLayer')._layer.removeLayer(this._layer);
+    this.get('parentComponent')._layer.removeLayer(this._layer);
   },
 
   leafletOptions: [],