+++ addon/components/tooltip-layer.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import DivOverlayLayer from 'ember-leaflet/components/div-overlay-layer';
 
-const { computed, run } = Ember;
+const { run, computed } = Ember;
 
 export default DivOverlayLayer.extend({
 
@@ -9,10 +9,8 @@ export default DivOverlayLayer.extend({
     'direction', 'permanent', 'sticky', 'interactive', 'opacity'
   ],
 
-  // creates a document fragment that will hold the DOM
-  destinationElement: computed(function() {
-    return document.createElement('div');
-  }),
+  // if this tooltip is permanent, we need to render the content immediately
+  shouldRender: computed.reads('permanent'),
 
   createLayer() {
     return this.L.tooltip(this.get('options')).setContent(this.get('destinationElement'));
@@ -27,11 +25,11 @@ export default DivOverlayLayer.extend({
   },
 
   addToContainer() {
-    this.get('containerLayer')._layer.bindTooltip(this._layer);
+    this.get('parentComponent')._layer.bindTooltip(this._layer);
   },
 
   removeFromContainer() {
-    this.get('containerLayer')._layer.unbindTooltip();
+    this.get('parentComponent')._layer.unbindTooltip();
   },
 
   _onLayerRemove({layer}) {
@@ -60,12 +58,12 @@ export default DivOverlayLayer.extend({
     };
     // we need to user `layerremove` event becase it's the only one that fires
     // *after* the popup was completely removed from the map
-    let containerLayer = this.get('containerLayer');
-    containerLayer._layer._map.addEventListener('layerremove', this._onLayerRemove, this);
+    let parentComponent = this.get('parentComponent');
+    parentComponent._layer._map.addEventListener('layerremove', this._onLayerRemove, this);
   },
 
   _removePopupListeners() {
-    let containerLayer = this.get('containerLayer');
-    containerLayer._layer._map.removeEventListener('layerremove', this._onLayerRemove, this);
+    let parentComponent = this.get('parentComponent');
+    parentComponent._layer._map.removeEventListener('layerremove', this._onLayerRemove, this);
   }
 });